+++ examples/complete/main.tf
@@ -4,8 +4,15 @@ provider "nomad" {}
 
 provider "vault" {}
 
+variable "nomad_jwks_url" {
+  description = "JWKS endpoint of the Nomad servers."
+  type        = string
+}
+
 module "workload_identity" {
   source = "../../modules/workload-identity"
+
+  nomad_jwks_url = var.nomad_jwks_url
 }
 
 module "traefik" {